refactor(agents): extract prompt-spec parsing from dispatch_plan god file
agent_task_dispatch_plan.rs was a 2209-line god file (audit god_file
finding). Extract the cohesive prompt/task-spec parsing cluster —
reading cook prompt/tasks-json inputs (raw strings, stored-prompt refs,
--tasks-json arrays) into validated DispatchPromptSpecs, resolving stored
prompts and deriving deterministic task ids — into a focused
prompt_spec.rs submodule.

Convert the file to a directory module (git mv preserves history). The
cluster had a clean seam: it references only stdlib/serde + agent_task_
prompts + config, no cross-cluster private state, and the plan builder
uses it through 5 functions + DispatchPromptSpec. Those are re-exported
pub(crate); the module's external pub contract (build_dispatch_plan,
build_dispatch_plan_with_provider_requirements,
preflight_dispatch_provider_secrets) is unchanged.

Behavior-preserving: mod.rs 2209 -> 1987 lines, prompt_spec.rs 239. All
33 dispatch_plan tests pass.
